fix: a lot of fixes for the bullet system
